# PoC: Use generics instead of code gen for vectors